basic VizController implemented but buggy
