+++ client/src/components/OpenLayers/MnCountySource.js
@@ -10,7 +10,7 @@ export default new VectorSource({
   url: function(extent) {
     return (
       "http://149.165.157.200:8080/geoserver/ows?service=wfs&" +
-      "version=1.1.0&request=GetFeature&typename=solap:mn_county_2010_1&" +
+      "version=1.1.0&request=GetFeature&typename=solap:mn_county_2010&" +
       "outputFormat=application/json&srsname=EPSG:3857"
     );
   },
